app/pyproject.toml
[project]
name = "daily-reporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "numpy>=2.2.1",
    "openai>=1.58.1",
    "openrouter>=1.0",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]
